requirements.txt
streamlit
pdfplumber
pymupdf
pandas
pytesseract
pillow
opencv-python
numpy
pdf2image
easyocr
transformers
torch
openpyxl
timm
